requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
openpyxl==3.1.2
# Integration and Security Features
django-otp==1.2.2
qrcode==7.4.2
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
